refactor: components 폴더 하위 폴더 ai_report 삭제
