+++ src/components/ai_report_sidebar.py
